+++ package.json
@@ -26,10 +26,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.25.1",
-<<<<<<< HEAD
     "react-otp-input": "^2.4.0",
-=======
->>>>>>> be0d0ee1f54e9edeea9a1029e9ebafd27a909088
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
