--- src/db/fakeAccount.ts
@@ -1,6 +1,0 @@
-import { User } from '../models/User';
-export const fakeAccount: User = {
-  citizenId: '123456789',
-  password: 'password123',
-  fullname: 'Phạm Công Thành'
-};
